+++ src/main/java/surveasy/domain/coupon/service/CouponService.java
@@ -1,43 +1,21 @@
 package surveasy.domain.coupon.service;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import surveasy.domain.coupon.dto.request.CouponCreateDTO;
 import surveasy.domain.coupon.dto.request.CouponUpdateDTO;
 import surveasy.domain.coupon.dto.response.CouponDiscountPercentResponse;
 import surveasy.domain.coupon.dto.response.CouponIdResponse;
 import surveasy.domain.coupon.dto.response.CouponListResponse;
-import surveasy.domain.coupon.helper.CouponHelper;
-import surveasy.domain.coupon.mapper.CouponMapper;
 
 @Service
-@RequiredArgsConstructor
-public class CouponService {
+public interface CouponService {
+    CouponDiscountPercentResponse getCouponDiscountPercent(String code);
 
-    private final CouponHelper couponHelper;
-    private final CouponMapper couponMapper;
+    CouponListResponse getCouponList();
 
-    public CouponDiscountPercentResponse getCouponDiscountPercent(String code) {
-        return couponMapper.toCouponDiscountPercentResponse(couponHelper.getCouponDiscountPercent(code));
-    }
+    CouponIdResponse createCoupon(CouponCreateDTO couponCreateDTO);
 
-    public CouponListResponse getCouponList() {
-        return couponMapper.toCouponListResponse(couponHelper.getCouponList());
-    }
+    CouponIdResponse updateCoupon(Long couponId, CouponUpdateDTO couponUpdateDTO);
 
-    @Transactional
-    public CouponIdResponse createCoupon(CouponCreateDTO couponCreateDTO) {
-        return couponMapper.toCouponIdResponse(couponHelper.createCoupon(couponCreateDTO));
-    }
-
-    @Transactional
-    public CouponIdResponse updateCoupon(Long couponId, CouponUpdateDTO couponUpdateDTO) {
-        return couponMapper.toCouponIdResponse(couponHelper.updateCoupon(couponId, couponUpdateDTO));
-    }
-
-    @Transactional
-    public void deleteCoupon(Long couponId) {
-        couponHelper.deleteCoupon(couponId);
-    }
+    void deleteCoupon(Long couponId);
 }
